requirements.txt
pandas
python-dotenv
supabase
google-generativeai
torch
sentence-transformers
streamlit